feat: user upcoming bookings on homepage (#361)
Display's a user's soonest booking on the homepage if said booking is today. If they hold multiple bookings for today, a button linking to their user page (where all bookings are listed) is also shown. Booking information allows a link-through to the venue and society for further information. The user can temporarily dismiss the information if they wish. 

---------

* initial logic and link to existing booking component

* query and componet data

* some name capitalisations

* button spacing

* component design

* component updates

* fix test

* Handle no query data better

* modified component and tests

* Fix lint:tsc errors

* Re-style stlightly and allow banner dismissal

* Fix test & type assertion

---------

Co-authored-by: Alex Toff <me@alextoff.uk>
Co-authored-by: HappyNTH <crayner.auburn@yahoo.co.uk>